fix: make C# returns summary optional
We're always adding the returns to summary when a method
was returning something, which could cause an error on
generating docs when return type is something like
`Task<ActionResult>`. In fact the `returns` from C# summary
should describe the return value instead of link the
return type, so this change is to make the returns on summary
optional.